+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-### [unreleased]
+## [unreleased]
+
+### Fixed
+
+- Handling of the internal `error` event within VrpcAgent.js to avoid an
+  exception in case the user forgot to implement that.
 
 ### Changed